epic: protobuf proposal
Proposal for adding detached extensions in the protobuf definitions.

We add an DetachedExtension to the PathSegmentExtensions.
This extension simply holds the digest of detached extensions.
The digest is a message such that we can add algorithms in the future
if they are necessary. The default will be sha256.

All extensions we do not yet to commit to, but we want to explore
are put in a separate experimental protobuf package. That package
is not covered by our backwards compatibility promise. When we have gained
some confidence in the extensions, and we want to commit to them,
they are moved to the real package.

The digest extension holds the value of 1000 such that we can reserve
it if we decide not to use the extension.