CHERRY-PICK: pkg/private/serrors: do not use reflect encoder for all context
Previously, the context was always encoded using the reflect encoder
(which is just json marshaling). This has the bad side effect that the
usual marshaling methods are not respected. E.g., error context that
would in theory implement the fmt.Stringer method were just JSON
encoded.

With this change, we leave it up to zap to decide how the values are
encoded based on their type.