CHERRY-PICK: control: check certificates available for beacons
Introduce a new beacon selection algorithm that ensures the certificate
chain for all the selected beacons is available locally. Beacons for
which this is not the case are not considered and not propagated.

This is required to counteract some edge cases where either the
certificate chain is lost (e.g., due to removing the trust database
manually) or if one of the hop entries has a shorter expiration time and
has expired in the meantime.