scion-pki: support additional extended key usage
Add the ability to specify additional extended key usage (EKU) in the
certificate creation process. This applies to `scion-pki certificate create`
and `scion-pki certificate sign` commands.

When setting the `--eku-any` flag is set, the ExtKeyUsageAny is set on
the certificate or CSR. Furthermore, the `--eku` flag allows the user
to provide a comma-separated list of ASN.1 OIDs for additional EKUs
that are unknown to scion-pki.